Updated my project

+++ src/main/java/com/revature/pn/reimbursements/NewReimbursementRequest.java
@@ -97,7 +97,7 @@ public class NewReimbursementRequest implements Request<Reimbursements> {
     public Reimbursements extractEntity() {
         Reimbursements extractEntity = new Reimbursements();
         extractEntity.setReimb_id(this.reimb_id);
-        extractEntity.setAmount(this.amount);
+        extractEntity.setAmount((float) this.amount);
         extractEntity.setSubmitted(this.submitted);
         extractEntity.setDescription(this.description);
         extractEntity.setAuthor_id(this.author_id);
